Move py tests to same package as src
